Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1197466 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/math/transform/FastHadamardTransformerTest.java
@@ -89,7 +89,7 @@ public final class FastHadamardTransformerTest {
         double dResult[] = transformer.transform(dX);
         for (int i = 0; i < dResult.length; i++) {
             // compare computed results to precomputed results
-            Assert.assertTrue(Precision.equals((double) y[i], dResult[i], 1));
+            Assert.assertTrue(Precision.equals(y[i], dResult[i], 1));
         }
     }
 
@@ -118,7 +118,7 @@ public final class FastHadamardTransformerTest {
         double dResult[] = transformer.inversetransform(dY);
         for (int i = 0; i < dResult.length; i++) {
             // compare computed results to precomputed results
-            Assert.assertTrue(Precision.equals((double) x[i], dResult[i], 1));
+            Assert.assertTrue(Precision.equals(x[i], dResult[i], 1));
         }
 
     }
